chore: clearing lint errors
