+++ components/Modals/DeleteScore.tsx
@@ -79,7 +79,7 @@ export default function DeleteScore({ open, setOpen, setSuccess, setFailure, sel
 								</div>
 								<div className='mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left'>
 									<Dialog.Title as='h3' className='text-lg leading-6 font-medium text-gray-900'>
-										Delete {selectedScore.user.first_name} {selectedScore.user.last_name}'s score for{' '}
+										Delete {selectedScore.user.first_name} {selectedScore.user.last_name}&apos;s score for{' '}
 										{selectedScore.date}?
 									</Dialog.Title>
 									<div className='mt-2'>
