+++ pages/user/[user].tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import Siderbar from '@/components/user/Sidebar';
 import UserHeader from '@/components/user/UserHeader';
 import { CogIcon, HomeIcon, PencilIcon } from '@heroicons/react/outline';
